igait_backend::daemons::filesystem
Function check_inputs_dir

async fn check_inputs_dir(entry: &DirEntry) -> Result<()>

Checks the a directory entry from the inputs folder, and preforms the following:

1.) Waits for a successful output

2.) Copies logfile to output directory (on Metis)

3.) Deletes the original logfile (on Metis)

4.) Deletes the local inputs folder

5.) Copies the output folder from Metis to the local outputs folder

Arguments

entry - The directory entry to check

Returns

A successful result if the directory was checked