Function igait_backend::daemons::filesystem::work_outputs

pub async fn work_outputs(app: Arc<Mutex<AppState>>)

The work inputs daemon, which checks the inputs directory for new jobs.

Arguments

app - The application state

Fails

If the inputs directory couldn’t be read

If the inputs directory couldn’t be iterated over

If the directory couldn’t be checked

Notes

This function never returns, ideally it should be run in a separate thread.