igait_backend::helper::email

Function _send_success_email

pub async fn _send_success_email(
    app: Arc<AppState>,
    recipient_email_address: &str,
    status: &JobStatus,
    datetime: &DateTime<Tz>,
    job: &Job,
    uid: &str,
    job_id: usize,
) -> Result<()>

A wrapper around send_email that sends a success email to the recipient.

Arguments

recipient_email_address - The email address to send the email to

status - The status of the job

datetime - The timestamp of the job

job - The job that was submitted

front_keyframed_url - The URL to the front keyframed video

side_keyframed_url - The URL to the side keyframed video

uid - The user ID of the job

job_id - The job ID of the job

Fails

If the email fails to send

Returns

A successful result if the email was sent

Notes

Any changes to the email logic should be made to the send_email function first