igait_backend::helper::email

Function send_contribution_email

pub async fn send_contribution_email(
    app: Arc<AppState>,
    email: &str,
    name: &str,
) -> Result<()>

A wrapper around send_email that thanks the recipient for their contribution.

Arguments

recipient_email_address - The email address to send the email to

status - The status of the job

datetime - The timestamp of the job

uid - The user ID of the job

job_id - The job ID of the job

name - The name entered by user from frontend

Fails

If the email fails to send

Returns

A successful result if the email was sent

Notes

Any changes to the email logic should be made to the send_email function first

The email is sent to the Cloudflare Worker at https://email-service.igaitniu.workers.dev/

The environment variable IGAIT_ACCESS_KEY is used to authenticate the request and must be set