Function igait_backend::helper::email::send_failure_email

pub async fn send_failure_email(
    recipient_email_address: &str,
    status: &JobStatus,
    dt_timestamp_utc: &DateTime<Utc>,
    uid: &str,
    job_id: usize,
    task_number: u128
) -> Result<()>

A wrapper around send_email that sends a failure email to the recipient.

Arguments

recipient_email_address - The email address to send the email to

status - The status of the job

dt_timestamp_utc - The timestamp of the job

uid - The user ID of the job

job_id - The job ID of the job

task_number - The task number of the job

Fails

If the email fails to send

Returns

A successful result if the email was sent

Notes

Any changes to the email logic should be made to the send_email function first