igait_backend::helper::email
Function send_welcome_email

pub async fn send_welcome_email(
    app: Arc<AppState>,
    job: &Job,
    uid: &str,
    job_id: usize,
) -> Result<()>

A wrapper around send_email that sends a welcome email to the recipient.

Arguments

job - The job that was submitted

uid - The user ID of the job

job_id - The job ID of the job

Fails

If the email fails to send

Returns

A successful result if the email was sent

Notes

Any changes to the email logic should be made to the send_email function first