igait_backend::helper
Module lib

Contains the filesystem helper functions and custom types.

Structs

AppError

The error type for the application.

AppState

AppStatePtr

Job

The job struct, which contains the job

JobStatus

The job status struct, which contains the status of the job.

Request

The request struct for the historical submissions endpoint.

UnauthorizedResponse

User

The user struct, which contains a user ID and a list of jobs.

Enums

JobStatusCode

The job status code enum, which contains the status codes for the job.

Functions

get_bearer_token 🔒