igait_backend::helper

Module metis

Contains the Metis helper functions.

Enums

SSHPath

Enum representing either a local or remote path, for use with the copy_file function

Constants

METIS_HOSTNAME

METIS_INPUTS_DIR

METIS_PBS_PATH

METIS_USERNAME

Functions

copy_file

A function that allows the user to copy files using the SCP command between one or two machines.

metis_qsub

Submits a PBS job to Metis via qsub

Type Aliases

PBSId