igait_backend
Module routes

This module contains the routes for the API.

To learn more about a route and how it works, click on the module name at the bottom of the page.

Modules

assistant

This module contains the assistant endpoint for the API.

contribute

This module contains the contribute endpoint for the API, which is used to submit a video for our research purposes.

historical

This module contains the historical submissions endpoint for the API.

upload

This module contains the job upload endpoint for the API.