igait_backend::routes
Module assistant

This module contains the assistant endpoint for the API.

Currently, this works by proxying requests to the iGait Assistant API.

This is done so due to the limitations of Firebase Auth expecting authentication via the Bearer header. Unfortunately, the WSS standard does not support this, so we instead extract it from an initial request and re-proxy to a now-authenticaed secure websocket. This approach is certainly not ideal, but it is indeed highly secure. In fact, it is significantly more secure than the WSS standard itself, which does not support authentication at all.

For more information on usage, see the iGAIT frontend codebase.

Structs

SearchJobArguments 🔒

Represents the arguments for the get_last_job and get_all_jobs functions.

Enums

AssistantUpdate 🔒

Represents an update from the assistant to the client.

Functions

assistant_entrypoint

The entry point for the assistant route.

assistant_proxied_entrypoint

The entry point for the API-facing assistant route.

handle_proxied_socket 🔒

Handles the proxied socket connection.

handle_proxied_socket_helper 🔒

A helper function to handle the proxied socket connection.

handle_socket 🔒

Handles the socket connection for the assistant route.

handle_socket_helper 🔒

A helper function to handle the socket connection.

send_response 🔒

Sends a response to the client over a mutable WebSocket connection.