igait_backend::routes::assistant
Function send_response

async fn send_response(
    app: &Arc<AppState>,
    client: &Client<OpenAIConfig>,
    thread: &ThreadObject,
    assistant: &AssistantObject,
    message: &str,
    socket: &mut WebSocket,
    user_id: &str,
) -> Result<()>

Sends a response to the client over a mutable WebSocket connection.

Arguments

app - The application state containing the OpenAI client and assistant.

client - The OpenAI client to use for sending the response.

thread - The thread object to send the response in.

assistant - The assistant object to use for the response.

message - The message to send to the assistant.

socket - The mutable WebSocket connection to send the response over.

user_id - The user ID of the client sending the message.

Fails

If the message fails to be created.

If the run fails to be created.

If the run fails to be polled.

If the run fails to be submitted.

If the message fails to be sent to the client.

If the run fails to be retrieved.

If the run fails to be submitted with tool outputs.

If the run fails to be retrieved after submitting tool outputs.

Returns

Nothing