igait_backend::routes
Module contribute

This module contains the contribute endpoint for the API, which is used to submit a video for our research purposes.

Arguments

uid: The Google Firebase user ID. Use the Firebase JS SDK to obtain before submitting.

email: The email address of the user of which to send emails to.

fileuploadfront: The front video file. Needs to be a file compatible with OpenPose.

fileuploadside: The side video file. Needs to be a file compatible with OpenPose.

name: The name of the user submitting the video.

Example cURL Request

curl -v -F fileuploadfront=@test.mp4 -F fileuploadside=test.mp4 /
    -F uid=curlplaceholder -F email=me@hiibolt.com -F name="John Doe" /
    http://api.igaitapp.com/api/v1/contribute

Potential Reasons for Failure

See each individual function for potential reasons for failure.

Structs

ContributeRequestArguments

A request to upload a video for the contribute endpoint.

ContributeRequestFile 🔒

A representation of a file in a Multipart request.

Functions

contribute_entrypoint

The entrypoint for the contribute request.

save_upload_files 🔒

Saves the upload files to S3 and the local filesystem.

unpack_contribute_arguments 🔒

Takes in the Multipart request and unpacks the arguments into a ContributeRequestArguments object.