igait_backend::routes::historical
Function get_email_and_pdf_link

async fn get_email_and_pdf_link(
    app: Arc<AppState>,
    jobs: Vec<(usize, Job)>,
    uid: String,
    timestamp: u64,
) -> Result<(String, String)>

Generates a PDF file containing the complete history of the user’s submissions.

Fails

If the PDF cannot be generated.

If the PDF cannot be read.

If the PDF cannot be uploaded to the user’s S3 bucket.

If the presigned URL cannot be generated.

Panics

If the PDF cannot be rendered.

Arguments

app - The application state.

jobs - The jobs to include in the PDF.

uid - The user ID.

timestamp - The timestamp to use for the PDF.

Returns

The email address of the user.

The presigned URL of the PDF.

Notes

This function is currently a bit of a hack. It uses a synchronous closure to generate the PDF.
It is possible for this function to panic without catching the panic.

Currently, I do not have the technical skills to fix this. I will come back with more skill later.