Module igait_backend::routes::upload

This module contains the job upload endpoint for the API.

Arguments

uid: The Google Firebase user ID. Use the Firebase JS SDK to obtain before submitting.

age: The age of the user in MM/DD/YYYY format.

ethnicity: The ethnicity of the user.

sex: The sex assigned to the patient at birth.

height: The height of the user in ft'in" format.

weight: The weight of the user in pounds.

email: The email address of the user of which to send emails to.

front_file: The front video file. Needs to be a file compatible with OpenPose.

side_file: The side video file. Needs to be a file compatible with OpenPose.

Example cURL Request

curl -v -F fileuploadfront=@test.mp4 -F fileuploadside=@test.mp4 /
    -F uid=curlplaceholder -F age=18 -F ethnicity=Caucasian /
    -F email=me@hiibolt.com -F sex=M -F height="5'10" -F weight=120 / 
    http://api.igaitapp.com/api/v1/upload

Potential Reasons for Failure

See each individual function for potential reasons for failure.

Notes

This route takes a long time to process. Make users aware of this on the frontend.

Structs

UploadRequestArguments 🔒

The required arguments for the upload request.

UploadRequestFile 🔒

A representation of a file in a Multipart request.

Functions

save_upload_files 🔒

Saves the upload files to S3 and the local filesystem.

unpack_upload_arguments 🔒

Takes in the Multipart request and unpacks the arguments into a UploadRequestArguments object.

upload_entrypoint

The entrypoint for the upload request.