igait_backend::routes::upload

Function unpack_upload_arguments

async fn unpack_upload_arguments(
    multipart: &mut Multipart,
) -> Result<UploadRequestArguments>

Takes in the Multipart request and unpacks the arguments into a UploadRequestArguments object.

Fails

If any of the fields are missing or if the files are too large.

Arguments

multipart - The Multipart object to unpack.