Module igait_backend::helper::metis

Contains the Metis helper functions.

Enums

SSHPath

Enum representing either a local or remote path, for use with the copy_file function

Constants

METIS_DATA_DIR

METIS_HOSTNAME

METIS_INPUTS_DIR

METIS_OUTPUTS_DIR

METIS_OUTPUT_NAME

METIS_PBS_PATH

METIS_USERNAME

Functions

copy_file

A function that allows the user to copy files using the SCP command between one or two machines.

delete_logfile

Removes a logfile off of Metis

delete_output_folder

Deletes an output folder off of Metis

metis_output_exists

Checks whether an output file exists on Metis

metis_qsub

Type Aliases

PBSId